fix: Dashboard stats, login UX, and redirect loop issues
DASHBOARD FIXES:
- Fixed Coordinator & Manager dashboard stats to show correct Rand amounts
- Changed TotalThisWeek from count to decimal for currency display
- Fixed ApprovedToday to track correct status changes using ClaimStatusHistory
- Updated queries to calculate weekly approved amounts accurately

LOGIN IMPROVEMENTS:
- Added test credentials panel with one-click auto-fill for all roles
- Added visible and functional Remember Me checkbox
- Implemented responsive two-column layout (credentials | form)
- Added visual feedback on credential selection

AUTHENTICATION FIXES:
- Fixed infinite redirect loop on app relaunch
- Added path checking in SessionAuthorizeAttribute to prevent circular redirects
- Updated cookie authentication with OnRedirectToLogin event handler
- Synced session timeout (8 hours) with auth cookie expiration

PDF GENERATION:
- Installed itext7.bouncy-castle-adapter (v8.0.5) for PDF report generation
- Fixed BouncyCastle dependency issue blocking report downloads

DATABASE:
- Fixed NULL constraint error for AdditionalNotes field in Claims table
- Added default empty string to prevent insertion failures

LAYOUT:
- Fixed jQuery loading issue preventing claim amount calculations
- Removed HTML comment hiding jQuery script tag in _Layout.cshtml

+++ CMCS/ViewModels/CoordinatorDashboardViewModel.cs
@@ -7,7 +7,10 @@ namespace CMCS.ViewModels
     {
         public List<Claim> ClaimsForReview { get; set; } = new List<Claim>();
         public int UrgentClaims { get; set; }
-        public int TotalThisWeek { get; set; }
+        
+        // ? FIXED: Changed to decimal for Rand amount
+        public decimal TotalThisWeek { get; set; }
+        
         public int PendingReview { get; set; }
         public int ApprovedToday { get; set; }
